backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0